refactor: White와 Black 중복 제거
