+++ src/main/java/chess/domain/piece/Piece.java
@@ -52,5 +52,5 @@ public abstract class Piece {
 
     public abstract double getScore();
 
-    public abstract void checkPawnMovable(Direction computeDirection, Piece targetPiece);
+    public abstract void checkPawnMovable(Direction direction, Piece targetPiece);
 }
