+++ src/main/java/chess/domain/state/Black.java
@@ -1,9 +1,7 @@
 package chess.domain.state;
 
 import chess.domain.board.Board;
-import chess.domain.board.Direction;
 import chess.domain.board.Location;
-import chess.domain.board.LocationDiff;
 import chess.domain.board.TeamScore;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Team;
@@ -15,14 +13,12 @@ public class Black extends Running {
 
     @Override
     public State move(Location source, Location target) {
-        checkMovable(source, target);
-
         Piece sourcePiece = getBoard().getPiece(source);
         Piece targetPiece = getBoard().getPiece(target);
-        LocationDiff locationDiff = source.computeDiff(target);
-        if (sourcePiece.isPawn()) {
-            sourcePiece.checkPawnMovable(locationDiff.computeDirection(), targetPiece);
-        }
+
+        super.checkMovable(source, target);
+        checkSourceColor(sourcePiece);
+        checkTarget(targetPiece);
 
         getBoard().move(source, target);
         if (targetPiece.isKing()) {
@@ -37,48 +33,14 @@ public class Black extends Running {
         return new TeamScore(Team.BLACK, score);
     }
 
-    private void checkMovable(Location source, Location target) {
-        Piece sourcePiece = getBoard().getPiece(source);
-        Piece targetPiece = getBoard().getPiece(target);
-        LocationDiff locationDiff = source.computeDiff(target);
-
-        checkSourceColor(sourcePiece);
-        checkDirection(sourcePiece, locationDiff.computeDirection());
-        checkDistance(sourcePiece, locationDiff);
-        checkRoute(source, locationDiff);
-        checkTarget(targetPiece);
-    }
-
-    private void checkSourceColor(Piece piece) {
-        if (!piece.isBlack()) {
+    private void checkSourceColor(Piece source) {
+        if (!source.isBlack()) {
             throw new IllegalArgumentException("[ERROR] 해당 말은 움직일 수 없습니다.");
         }
     }
 
-    private void checkDirection(Piece piece, Direction direction) {
-        if (!piece.isMovableDirection(direction)) {
-            throw new IllegalArgumentException("[ERROR] 해당 위치로 이동할 수 없습니다.");
-        }
-    }
-
-    private void checkDistance(Piece piece, LocationDiff locationDiff) {
-        if (!piece.isMovableDistance(locationDiff)) {
-            throw new IllegalArgumentException("[ERROR] 해당 위치까지 이동할 수 없습니다.");
-        }
-    }
-
-    private void checkRoute(Location source, LocationDiff locationDiff) {
-        Location routeLocation = source.copyOf();
-        for (int i = 0; i < locationDiff.computeDistance() - 1; i++) {
-            routeLocation = routeLocation.add(locationDiff.computeDirection());
-            if (!getBoard().isEmpty(routeLocation)) {
-                throw new IllegalArgumentException("[ERROR] 해당 경로를 지나갈 수 없습니다. ");
-            }
-        }
-    }
-
-    private void checkTarget(Piece targetPiece) {
-        if (targetPiece.isBlack()) {
+    private void checkTarget(Piece target) {
+        if (target.isBlack()) {
             throw new IllegalArgumentException("[ERROR] 같은 색의 기물을 잡을 수 없습니다.");
         }
     }
